Overhaul Cel, replace compile-time ('ct') model with skeletons, move to Java 8.
The Cel overhaul also implies that @Require, @Ensure and @Invariant have been replaced with a single annotation @Contract.
Also, the @Equivalent annotation has been replaced with the annotations in ch.raffael.contracts.macro.
Upgrade ANTLR to version 4.2.2.
The documentation for Cel still needs to be updated, it doesn't currently reflect the actual Cel.